Create extract/catalog.ts as per @viceice's suggestion
