+++ lib/manager/gradle/shallow/extract.ts
@@ -1,4 +1,3 @@
-import { parse } from '@iarna/toml';
 import * as upath from 'upath';
 import {
   id as datasource,
@@ -12,13 +11,9 @@ import type {
   PackageFile,
 } from '../../types';
 import type { GradleManagerData } from '../types';
+import { parseCatalog } from './extract/catalog';
 import { parseGradle, parseProps } from './parser';
-import type {
-  GradleCatalog,
-  GradleCatalogPluginDescriptor,
-  PackageVariables,
-  VariableRegistry,
-} from './types';
+import type { PackageVariables, VariableRegistry } from './types';
 import {
   getVars,
   isGradleFile,
@@ -81,69 +76,9 @@ export async function extractAllPackageFiles(
         updateVars(vars);
         extractedDeps.push(...deps);
       } else if (isTOMLFile(packageFile)) {
-        // Implement TOML file parsing and extraction
-        const tomlContent = parse(content) as GradleCatalog;
-        const versions = tomlContent.versions || {};
-        const libs = tomlContent.libraries || {};
-        const libStartIndex = content.indexOf('libraries');
-        const libSubContent = content.slice(libStartIndex);
-        const versionStartIndex = content.indexOf('versions');
-        const versionSubContent = content.slice(versionStartIndex);
-        for (const libraryName of Object.keys(libs)) {
-          const libDescriptor = libs[libraryName];
-          const group: string =
-            typeof libDescriptor === 'string'
-              ? libDescriptor.split(':')[0]
-              : libDescriptor.group || libDescriptor.module?.split(':')[0];
-          const name: string =
-            typeof libDescriptor === 'string'
-              ? libDescriptor.split(':')[1]
-              : libDescriptor.name || libDescriptor.module?.split(':')[1];
-          const version = libDescriptor.version || libDescriptor.split(':')[2];
-          const currentVersion =
-            typeof version === 'string' ? version : versions[version.ref];
-          const fileReplacePosition =
-            typeof version === 'string'
-              ? libStartIndex +
-                findIndexAfter(libSubContent, libraryName, currentVersion)
-              : versionStartIndex +
-                findIndexAfter(versionSubContent, version.ref, currentVersion);
-          const dependency = {
-            depName: `${group}:${name}`,
-            groupName: group,
-            currentValue: currentVersion,
-            managerData: { fileReplacePosition, packageFile },
-          };
-          extractedDeps.push(dependency);
-        }
-        const plugins = tomlContent.plugins || {};
-        const pluginsStartIndex = content.indexOf('[plugins]');
-        const pluginsSubContent = content.slice(pluginsStartIndex);
-        for (const pluginName of Object.keys(plugins)) {
-          const pluginDescriptor = plugins[
-            pluginName
-          ] as GradleCatalogPluginDescriptor;
-          const pluginId = pluginDescriptor.id;
-          const version = pluginDescriptor.version;
-          const currentVersion: string =
-            typeof version === 'string' ? version : versions[version.ref];
-          const fileReplacePosition =
-            typeof version === 'string'
-              ? pluginsStartIndex +
-                findIndexAfter(pluginsSubContent, pluginId, currentVersion)
-              : versionStartIndex +
-                findIndexAfter(versionSubContent, version.ref, currentVersion);
-          const dependency = {
-            depType: 'plugin',
-            depName: pluginId,
-            lookupName: `${pluginId}:${pluginId}.gradle.plugin`,
-            registryUrls: ['https://plugins.gradle.org/m2/'],
-            currentValue: currentVersion,
-            commitMessageTopic: `plugin ${pluginName}`,
-            managerData: { fileReplacePosition, packageFile },
-          };
-          extractedDeps.push(dependency);
-        }
+        const updatesFromCatalog = parseCatalog(packageFile, content);
+        console.log(updatesFromCatalog);
+        extractedDeps.push(...updatesFromCatalog);
       } else if (isGradleFile(packageFile)) {
         const vars = getVars(registry, dir);
         const {
